WIP: prisma seed is segfaulting again 😭.
