+++ participant/Dockerfile
@@ -9,7 +9,10 @@ RUN apk update \
     # Install curl for healthcheck
     curl \
     # Install openssl for Prisma
-    openssl
+    openssl \
+    # Install for debug
+    libunwind-dev \
+    python3 g++ make
 # Set working directory and node user
 WORKDIR /home/node/remix
 
@@ -26,7 +29,7 @@ COPY --from=deps /home/node/remix/node_modules ./node_modules
 COPY --chown=node:node prisma .
 RUN npx prisma generate
 COPY --chown=node:node . .
-RUN cp ./prisma/seed.not-prod.ts ./prisma/seed.ts
+RUN cp public/data/submissions.not-prod.json public/data/submissions.json
 CMD ["npm", "run", "dev:docker"]
 
 # -------- Image: e2e -------- #
@@ -65,6 +68,5 @@ COPY --from=build /home/node/remix/node_modules/.prisma ./node_modules/.prisma
 COPY --from=build /home/node/remix/build ./build
 COPY --from=build /home/node/remix/public/build ./public/build
 COPY --chown=node:node . .
-RUN cp ./prisma/seed.prod.ts ./prisma/seed.ts
 VOLUME ["/tmp"]
 CMD ["/home/node/remix/bin/prod-start.sh"]
