+++ participant/package.json
@@ -35,7 +35,7 @@
     "/public/build"
   ],
   "dependencies": {
-    "@prisma/client": "^4.5.0",
+    "@prisma/client": "^4.12.0-integration-rtld-deepbind.3",
     "@remix-run/express": "^1.7.6",
     "@remix-run/node": "^1.7.6",
     "@remix-run/react": "^1.7.6",
@@ -104,12 +104,13 @@
     "jest-environment-jsdom": "^29.3.1",
     "jest-mock-extended": "^3.0.1",
     "js-image-generator": "^1.0.4",
+    "node-segfault-handler": "^1.4.2",
     "npm-run-all": "^4.1.5",
     "playwright-core": "^1.30.0",
     "postcss": "^8.4.21",
     "postcss-cli": "^10.1.0",
     "prettier": "^2.8.4",
-    "prisma": "^4.10.1",
+    "prisma": "^4.12.0-integration-rtld-deepbind.3",
     "sass": "^1.56.1",
     "tmp": "^0.2.1",
     "ts-jest": "^29.0.3",
