+++ participant/public/data/submissions.json
@@ -1,0 +1,1 @@
+{}
